feat: add snapshot based AST testing
