+++ Cargo.toml
@@ -24,6 +24,7 @@ colored = "2.0.0"
 glob = "0.3"
 heraclitus-compiler = "1.8.1"
 include_dir = "0.7.4"
+insta = "1.43.1"
 itertools = "0.13.0"
 similar-string = "1.4.2"
 test-generator = "0.3.1"
