+++ src/tests/translation/add_int_expression.ab
@@ -1,0 +1,3 @@
+echo 15 + 45
+echo 39 + 21 + 80
+echo 3 + 9 + 648 + 1232
