+++ src/tests/translation/add_int_shorthand.ab
@@ -1,0 +1,4 @@
+let x = 10
+x += 5
+let y = 20
+y += 15
