+++ src/tests/translation/add_num_expression.ab
@@ -1,0 +1,3 @@
+echo 15.15 + 45.15
+echo 39.45 + 21.18 + 80.12
+echo 3.0 + 9.0 + 648.0 + 12.23
