+++ src/tests/translation/add_num_shorthand.ab
@@ -1,0 +1,4 @@
+let x = 10.5
+x += 5.2
+let y = 20.1
+y += 15.7
