+++ src/tests/translation/div_int_expression.ab
@@ -1,0 +1,3 @@
+echo 10 / 2
+echo 20 / 4 / 2
+echo 100 / 5 / 2 / 2
