+++ src/tests/translation/div_int_shorthand.ab
@@ -1,0 +1,4 @@
+let x = 20
+x /= 4
+let y = 35
+y /= 5
