+++ src/tests/translation/div_num_expression.ab
@@ -1,0 +1,3 @@
+echo 10.5 / 2.1
+echo 20.4 / 4.0 / 2.0
+echo 100.0 / 5.0 / 2.5 / 2.0
