+++ src/tests/translation/div_num_shorthand.ab
@@ -1,0 +1,4 @@
+let x = 20.8
+x /= 4.2
+let y = 35.5
+y /= 5.0
