+++ src/tests/translation/modulo_int_expression.ab
@@ -1,0 +1,3 @@
+echo 10 % 3
+echo 20 % 6 % 3
+echo 100 % 7 % 3 % 2
