+++ src/tests/translation/modulo_int_shorthand.ab
@@ -1,0 +1,4 @@
+let x = 20
+x %= 7
+let y = 35
+y %= 6
