+++ src/tests/translation/modulo_num_expression.ab
@@ -1,0 +1,3 @@
+echo 10.5 % 3.2
+echo 20.4 % 6.0 % 3.1
+echo 100.0 % 7.5 % 3.0 % 2.5
