+++ src/tests/translation/modulo_num_shorthand.ab
@@ -1,0 +1,4 @@
+let x = 20.5
+x %= 7.2
+let y = 35.8
+y %= 6.3
