+++ src/tests/translation/mul_int_expression.ab
@@ -1,0 +1,3 @@
+echo 5 * 3
+echo 2 * 4 * 6
+echo 1 * 2 * 3 * 4
