+++ src/tests/translation/mul_int_shorthand.ab
@@ -1,0 +1,4 @@
+let x = 5
+x *= 4
+let y = 7
+y *= 6
