+++ src/tests/translation/mul_num_expression.ab
@@ -1,0 +1,3 @@
+echo 5.5 * 3.2
+echo 2.1 * 4.0 * 6.3
+echo 1.0 * 2.5 * 3.1 * 4.7
