+++ src/tests/translation/mul_num_shorthand.ab
@@ -1,0 +1,4 @@
+let x = 5.5
+x *= 4.2
+let y = 7.1
+y *= 6.8
