+++ src/tests/translation/sub_int_expression.ab
@@ -1,0 +1,3 @@
+echo 21 - 7
+echo 2 - 1 - 3
+echo 1 - 3 - 2 - 12
