+++ src/tests/translation/sub_int_shorthand.ab
@@ -1,0 +1,4 @@
+let x = 20
+x -= 8
+let y = 50
+y -= 25
