+++ src/tests/translation/sub_num_expression.ab
@@ -1,0 +1,3 @@
+echo 21.2 - 7.3
+echo 2.32 - 1.12 - 3.32
+echo 1.5 - 3.5 - 2.5
