+++ src/tests/translation/sub_num_shorthand.ab
@@ -1,0 +1,4 @@
+let x = 20.5
+x -= 8.2
+let y = 50.8
+y -= 25.3
